requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
alembic=1.11.3=pypi_0
blas=1.0=mkl
bottleneck=1.3.7=pypi_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.7.22=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cmaes=0.10.0=pypi_0
colorlog=6.7.0=pypi_0
cryptography=41.0.2=py310h22a60cf_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.10.0=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
greenlet=2.0.2=pypi_0
huggingface-hub=0.17.3=pypi_0
idna=3.4=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py310h06a4308_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
mako=1.2.4=pypi_0
markupsafe=2.1.1=py310h7f8727e_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.6=py310h1128e8f_1
mkl_random=1.2.2=py310h1128e8f_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
numpy=1.25.2=py310h5f9d8c6_0
numpy-base=1.25.2=py310hb5e798b_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.10=h7f8727e_2
optuna=3.3.0=pypi_0
packaging=23.1=pypi_0
pillow=9.4.0=py310h6a678d5_0
pip=23.2.1=py310h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.12=h955ad1f_0
pytorch=2.0.1=py3.10_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=py310h06a4308_0
safetensors=0.4.0=pypi_0
scipy=1.11.2=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.0.0=py310h06a4308_0
sqlalchemy=2.0.20=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.14.1=pypi_0
torchaudio=2.0.2=py310_cu117
torchtriton=2.0.0=py310
torchvision=0.15.2=py310_cu117
tqdm=4.66.1=pypi_0
transformers=4.35.0=pypi_0
typing_extensions=4.7.1=py310h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py310h06a4308_0
wheel=0.38.4=py310h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
